\name{print.ivregDecision}
\alias{print.ivregDecision}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Printing object of class "ivregDecision"
}
\description{
	"ivregDecisionList", a list of objects of class "ivregDecision", is part 
	of an output of sensitivity analysis using function \code{'sensIvreg'}.
}
\usage{
\S3method{print}{ivregDecision}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
	object of class "ivregDecision".
}
	\item{\dots}{Ignore.}
}
\author{
Bikram Karmakar
}
\note{
Trivially applies to object of class "ivregDecisionList" 
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{plot.ivregDecision}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ Sensitivity Analysis }

